# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task API",
    "description": "A simple REST API for managing tasks built with Express and TypeScript.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "tags": [
    {
      "name": "General",
      "description": "General API endpoints"
    },
    {
      "name": "Tasks",
      "description": "Task CRUD operations"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": ["General"],
        "summary": "API information",
        "description": "Returns basic information about the API.",
        "responses": {
          "200": {
            "description": "API information returned successfully."
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": ["General"],
        "summary": "Health check",
        "description": "Checks whether the API is running.",
        "responses": {
          "200": {
            "description": "API is healthy."
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get all tasks",
        "description": "Returns every task in the application.",
        "responses": {
          "200": {
            "description": "A list of tasks.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tasks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Task"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tasks"],
        "summary": "Create a new task",
        "description": "Creates a new task using the supplied title.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTask"
              },
              "example": {
                "title": "Learn Swagger"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input."
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get a task by ID",
        "description": "Returns a single task.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "Task not found."
          }
        }
      },
      "put": {
        "tags": ["Tasks"],
        "summary": "Update a task",
        "description": "Updates the title and completion status of a task.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTask"
              },
              "example": {
                "title": "Master Swagger UI",
                "done": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body."
          },
          "404": {
            "description": "Task not found."
          }
        }
      },
      "delete": {
        "tags": ["Tasks"],
        "summary": "Delete a task",
        "description": "Deletes a task by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task deleted successfully."
          },
          "404": {
            "description": "Task not found."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Learn Express"
          },
          "done": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "CreateTask": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Learn OpenAPI"
          }
        }
      },
      "UpdateTask": {
        "type": "object",
        "required": ["title", "done"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Learn Swagger UI"
          },
          "done": {
            "type": "boolean",
            "example": true
          }
        }
      }
    }
  }
}